% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmeans.R
\name{group_means}
\alias{group_means}
\title{Calculate Group Means}
\usage{
group_means(dataset, group_cols)
}
\arguments{
\item{dataset}{A data frame containing the data to be analyzed. The choice of
\code{dataset} as a parameter name reflects its role as the primary
input for the function, representing the data source.}

\item{group_cols}{A character vector of column names by which to group the data.
This name was chosen to clearly indicate that it represents
the columns that define the groups for the mean calculation.}
}
\value{
A data frame containing the grouped means of all numeric columns, with
one row per unique combination of the specified grouping columns.
}
\description{
This function calculates the mean of all numeric columns in a given dataset,
grouped by specified categorical columns.
}
\examples{
df <- data.frame(
  group1 = c("A", "A", "B", "B", "A", "B"),
  group2 = c("X", "Y", "X", "Y", "X", "Y"),
  value1 = c(10, 20, 30, 40, 20 ,30),
  value2 = c(5, 15, 25, 35, 5, 15)
)
group_means(df, c("group1", "group2"))

}
